move lazy inputfield helper to common extension
